frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "The frontend for JobScout"
authors = [
    {name = "Mateusz Piechowski",email = "you@example.com"},
    {name = "Mariusz Rychlicki",email = "mariusz.rychlicki@proton.me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "nicegui>=1.4.0",
    "httpx>=0.28.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
